Refactor local connector option loading
