+++ cli/connections/bigQuery.ts
@@ -1,4 +1,5 @@
 import {BigQuery, BigQueryDate, BigQueryTimestamp, type BigQueryOptions} from '@google-cloud/bigquery'
+import {readFileSync} from 'fs'
 
 import {config} from '../../lang/config.ts'
 import {type QueryConnection, type QueryResult, type SchemaColumn, type QueryOptions} from './types.ts'
@@ -87,4 +88,16 @@ export function normalizeBigQueryRows(rows: Record<string, any>[]) {
       if (v instanceof BigQueryDate) r[k] = v.value
     })
   })
+}
+
+export function localDbOptions(): BigQueryOptions {
+  if (process.env.GOOGLE_CREDENTIALS_CONTENT) {
+    let credentials = JSON.parse(process.env.GOOGLE_CREDENTIALS_CONTENT)
+    return {projectId: credentials.project_id, credentials}
+  }
+  if (process.env.GOOGLE_APPLICATION_CREDENTIALS) {
+    let credentials = JSON.parse(readFileSync(process.env.GOOGLE_APPLICATION_CREDENTIALS, {encoding: 'utf-8'}))
+    return {projectId: credentials.project_id}
+  }
+  return {}
 }
